platform/qemu: boot ISO as SCSI CD on s390x
Using the `-drive` shortcut doesn't work on s390x. Be more explicit by
specifying the block device and the SCSI CD drive. This matches
https://www.qemu.org/docs/master/system/s390x/bootdevices.html.

At the QMP level, this empties the `device` key we previously relied on
to find the CD-ROM boot device when changing boot order. Adapt it by
having it look at the `node-name` key instead.