kola/testiso: add ISO tests on s390x
Add `iso-live-login` which should already work.

Now that we have Ignition embedding support in the s390x ISO, we can
also test `iso-offline-install`. `iso-install` should also work but for
consistency with other !x86_64 arches, we don't test it.

The `mpath` variant is currently failing but should work. Keep it
disabled for now so we don't introduce a failure that immediately has to
be denylisted.