use new bucket layout for release stream metadata
The bucket layout was changed when migrating the RHCOS pipeline to
use the same code as the FCOS pipeline. Using one instance of Jenkins
allowed all arches to use the same build ID and therefore share the
parent build directory.

Plume is used to update the boot images in the OCP installer so update
it to handl the new bucket layout.