platform/qemu: report clearer error when device not found
Right now, we don't check if either of `bootdev` or `primarydev` isn't
found, so we go and tell QEMU over QMP to change the boot order for
device `""`, which results in a confusing QEMU error message.

Let's check for these error conditions upfront and also include the full
list of devices we scanned.