cmd-fetch: support "autolocking" from an arch build
In a multi-arch build, we want to be able to provide some guarantees
that the RPM set across arches is the same (of course, save for any
arch-specific packages). The canonical way to do this is with lockfiles,
but RHCOS currently doesn't have that. However, since we currently build
x86_64 before building the multi-arch variants, what we can do is derive
a lockfile from the x86_64 build and use that in the multi-arch builds.
This patch adds a new "autolock" feature to the `fetch` and `build`
commands to do just that.

Even once RHCOS gets lockfiles, this will also be useful for development
streams where we purposely don't want to maintain lockfiles (such as
rawhide).

This isn't foolproof. Notably if a package is included in more than one
multi-arch build, but not x86_64, it's possible for that package to
diverge. Presently in RHCOS 4.13 at least, only one such package exists
(`libatomic`), so in practice that risk is quite low.