Merge mantle into toplevel module
Having separate modules for the toplevel cosa and mantle is confusing
and inefficient. Let's finish the mantle merger we started long ago and
make it part of cosa proper. This means dropping the mantle go module
and having it be a package of the toplevel coreos-assembler one.

This allows us to have a single `vendor/` directory and Makefile. This
also makes schema updates easier because we're no longer vendoring it
for the sake of mantle. Another nice thing this fixes is that the golang
language server (`gopls`) is now no longer confused by how our repo is
set up and works nicely OOTB.

Follow-up patches will merge the `vendor/` directories and Makefiles.
In this patch, we only update package paths. We could in the future
hoist bits from `mantle/` to the toplevel, but that's much more invasive
and will break some permalinks so leave it alone for now.

The end result is that the mantle path is changing from
`github.com/coreos/mantle` to
`github.com/coreos/coreos-assembler/mantle`.

This in theory breaks any external user trying to use mantle packages,
but AFAIK, we don't provide any public APIs in this repo.

To do this change, I ran the following commands:

```
$ git grep --name-only \"github.com/coreos/mantle\" | \
  xargs sed -i -e 's,"github.com/coreos/mantle","github.com/coreos/coreos-assembler/mantle",'
$ git grep --name-only \"github.com/coreos/mantle/.*\" | \
  xargs sed -i -e 's,"github.com/coreos/mantle/\(.*\)","github.com/coreos/coreos-assembler/mantle/\1",'
```

The escaped quotes matched nicely the import and e.g. logging instances
in the codebase. I updated a few remaining ones by hand. There are a
bunch of dead links to the github.com/coreos/mantle repo that predate
this PR that I left alone for now.