cmd-init: add support for `--yumrepos-branch`
In the usual case, we don't need to change the branch of the yumrepos
repo since for RHCOS (the only current user), all repos for all releases
are located in the default branch.

However, while reviewing the hotfix process with ART, we realized that
it would be useful to support this so that we can have the yum repo
definitions live in the same git repo and branch as the hotfix pipecfg
definition itself.