kola: Add `skip-console-warnings` for `kola-denylist.yaml`
We have a case in https://bugzilla.redhat.com/show_bug.cgi?id=2164765
where we want to for now ignore a kernel warning across *all* tests.

Add a special case pattern `skip-console-warnings` that implements
this.

I tested this by doing:

```
echo '<0>Oops: synthetic' > /dev/kmsg
```
in a test, and verified that the test fails without the pattern
set, and succeeds with it.