mantle/platform/aws: ignore local and wavelength availability zones
Aashish was getting an error when trying to run kola in a different AWS
account where it was complaining about an incorrect IPv6 CIDR:

```
2023-01-30T20:50:05Z cli: Spawning instance failed: error resolving
security group: creating subnets: creating subnet: InvalidSubnet.Range:
The IPv6 CIDR '2600:1f18:edb:e906::/64' is invalid.
```

After some inspection it was apparent that he was seeing more security
groups than I was in the `us-east-1` region. It turns out this is
because local and wavelength availability zones exist [1].

I'm assuming the IPv6 CIDR error message is actually not the problem but
more that a local zone exists that doesn't support IPv6. This is an
untested theory.

Either way, let's just ignore non primary availability zones for now.

[1] https://aws.amazon.com/about-aws/global-infrastructure/localzones/locations