mantle: overhaul azure authentication
This is a partial backport of 50f6205. Rather than doing any
library changes here (since the versions of libraries on these
older branches are mostly frozen in time) we are just picking up
the changes to the credentials file we use as input where we now
have a single azureCreds.json rather than two (azureProfile.json
and azureAuth.json).

In order to support this I still need to get the information back
into a AZURE_AUTH_LOCATION formatted file for the old library to be
able to read the data. For now we'll just write it out to a temporary
file that gets cleaned up immediately after use. This isn't great,
but the limited usage here only for older branches executed in our
downstream pipeline makes it less of a concern.