mantle/platform/api: overhaul Azure storage API code
The github.com/Microsoft/azure-vhd-utils code hasn't had a commit since
2018 and I'm not sure it was ever intended to be used as a library (i.e.
there's a lot of `go mod vendor` that won't work right anymore because
it's so out of date and I'm not sure that's viewed as a bad thing).

Either way let's get rid of the old storage code and write new code
using the new sdk libraries for interacting with blob storage.