create_disk: Create image layer refs by default
Fedora CoreOS is not yet using containers by default for updates;
xref https://github.com/coreos/fedora-coreos-tracker/issues/1263
etc.

Consequently, when one boots a FCOS system and wants
to rebase to a custom image, one ends up downloading the entire image,
including the parts of FCOS that you already have.

This changes things so that when we generate disk images by default,
we write the *layer refs* of the component parts - but we
delete the "merged" container image ref.

The semantics here will be:

- Only a tiny amount of additional data used by default;
  the layer refs are just metadata, the bulk of the data still
  lives in regular file content.
- When a FCOS system auto-updates to its by-default usage of
  an ostree commit, the unused layer refs will be garbage collected.
- But, as noted above when rebasing to a container image instead,
  if the target container image reuses some of those layers (as
  we expect when rebasing FCOS to a FCOS-derived container) then
  we don't need to redownload them - we only download what the user
  provided.

Hence, this significantly improves rebasing to container images,
with basically no downsides.

The alternative code path to actually deploy *as a container*
remains off by default.  When that is enabled, `rpm-ostree upgrade`
fetches a container by default, which is a distinct thing.