Refactor buildextend compression and add zip
buildextend --compress supports gzip as an alternate compression over
xz using a key called "gzip" where the value is a bool.  This is now
refactored to a key called "compression" with a value of "gzip", "skip",
or "zip" (see below).  The "skip-compression" key is no longer
applicable.

HyperV images which are used on Windows need to be compressed with zip
because xz and gzip are not supported natively.  Added zip support to
buildextend and set the hyperv platform compression to "zip".

Supports:
  * https://github.com/coreos/fedora-coreos-tracker/issues/1411
  * https://github.com/coreos/fedora-coreos-tracker/issues/1424

Signed-off-by: Brent Baude <bbaude@redhat.com>