buildextend-hashlist-experimental: make tmp checkout readable
We are changing the 'find' invocation permissions to access the
/etc/credstore directory.'find' can't recurse into '/etc/credstore',
we add 'x' so that it can recurse into it. With '+', 'find' accumulates
paths and runs 'chmod' once (or batched by the arg limit). But for
this, we need it to 'chmod' the dir before it tries to recurse into
it, so we change to ';'.

Fixes: coreos/fedora-coreos-tracker#1429