mantle/aws: set boot mode to `uefi-preferred`
AWS now allows marking an AMI as "UEFI-preferred"; i.e. that both BIOS
and UEFI are supported, but the latter is preferred if the instance type
supports it.

Let's make use of it.

A change proposal has been submitted for Fedora 39 to also do this for
the Fedora Cloud image.

Ref: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ami-boot.html
Ref: https://fedoraproject.org/wiki/Changes/CloudEC2UEFIPreferred