mantle/platform/conf: use 3.1.0 Ignition spec version for MergeAllConfigs()
For MergeAllConfigs() let's use an older spec version to enable testing
bringup on older Fedora CoreOS releases in the aid of extended upgrade
testing. Specifically, now the --append-butane workflow enabled and
described in [1] will work.

I chose spec 3.1.0 because it was old enough (F32 nodes support it)
and it includes [2], which enabled us to not change using the Resource()
type in Ignition, which didn't exist in 3.0.0.

[1] https://github.com/coreos/fedora-coreos-config/commit/2b0fe09b75c7200b538724be97081f035085164d
[2] https://github.com/coreos/ignition/commit/082b45478d18381f959ec578846a19366344268e