mantle/platform/conf: transparently handle Butane spec stabilizations
There are three stabilization cases we need to handle:

1. Tests using an experimental Ignition spec that's being stabilized.
2. Tests using an experimental Butane spec that's being stabilized, where
   the Butane spec already produces a stable Ignition spec.
3. Tests using an experimental Butane spec that's being stabilized, where
   the Butane spec produces an Ignition spec which is also being
   stabilized.

We already have a procedure for case 1.  Ignition configs in tests are
stabilized by the PR that lands the new Ignition in
fedora-coreos-config, and mantle has code to detect the next stable
Ignition spec (otherwise unparsable) and wrap it in an older spec
version that mantle does know how to handle.  This is good enough for
Ignition in the image, and avoids requiring immediate revendoring of
the new Ignition into mantle.

Implement a similar approach for cases 2 and 3.  Butane configs in
tests can be stabilized by the PR that updates Ignition in the image.
When we see a Butane config with an unrecognized variant/version pair,
try stapling "-experimental" onto the spec version and see if we can
parse the config.  That's sufficient for case 2, but case 3 has the
additional complication that Butane's output will no longer be accepted
by Ignition in the image.  Handle that by stripping "-experimental"
from the Ignition config version.  (Parsing the result will succeed
because of the case 1 workaround.)

As a result, revendoring of both Ignition and Butane into mantle can be
deferred until convenient, allowing both projects to be updated
simultaneously once the new Butane release is out.  This is important
because there's generally a delay between the two releases, and the
previous Butane version will depend on API that's missing from the new
Ignition version.

Fixes https://github.com/coreos/butane/issues/451.