mantle/platform/conf: use 3.0.0 Ignition spec version for MergeAllConfigs()
This lowers the spec version to 3.0.0, which is the minimum version
we're guaranteed to support so it works nicely as a "head config". A
minor bonus is that it allows us to drop a workaround for it in the
upgrade tests pipeline code.[[1]]

Follow-up to d1e794ee4 ("mantle/platform/conf: use 3.1.0 Ignition spec
version for MergeAllConfigs()").

[1]: https://github.com/coreos/fedora-coreos-pipeline/pull/853