cmd-remote-build-container: query single tag instead of all tags
When checking if a tag exists in a repo, we were using
`podman search --list-tags`. However, by default it limits the number of
tags to 25.

We could bump the limit, but rather than do that let's just switch
to `skopeo inspect` to query the single tag since querying tags is an
expensive operation. The downside is that it's somewhat hacky to handle
the `ENOENT` case.