platform/metal: set serial console baud rate on x86_64
Otherwise, the default is a ridiculously small rate. Do the same in
`cmdlib`, from which I think we originally copied this table.

Note these args are used purely for VMs we bring up for building
and testing things. The kargs in shipped artifacts are controlled by
`platforms.yaml` in the source config. This patch makes the console karg
used by cosa itself match what we have in `platforms.yaml`.