cmd-generate-release-meta: fix kubevirt object in release metadata
We're not currently setting all the fields of the `kubevirt` image in
release metadata. Notably as per earlier discussions[[1]][[2]], we want
the `image` field to use a floating tag and the `digest-ref` field to
use a digest pullspec.

Determining the floating tag to use is a bit hacky right now, but safe:
we find in the list of pushed tags the one which doesn't have a build ID
reference to it. If we find multiple, we fail. In the future, we'll
probably want to strengthen this to be more explicit.

[1]: https://github.com/coreos/stream-metadata-go/pull/46
[2]: https://github.com/coreos/fedora-coreos-tracker/pull/1172

Closes: #3456