cmd-init: run `git submodule update` when `--commit` is used
When building RHCOS, we first build for x86_64 on the latest commit
available on the stream's branch. When we trigger builds for other
arches, we want to make sure the same commit of the src config is used,
so we pass `--commit`.

However, `--commit` doesn't reset git submodule, so there's a race
condition where if a submodule bump landed between the x86_64 build and
the multi-arch builds, we would use the latest submodule, rather than
the older one.

Run `git submodule update` to fix this race.