kola/test: update rhcos upgrade test to new s3 dir
Originally RHCOS only had one stream named after the OCP major minor
version. Around the 4.12 timeframe, the ability to produce concurrent
RHCOS versions based on different RHEL content was introduced. This
changed the stream name used in the mirror where this test downloads
the latest released RHCOS version.

Since the stream name is not in the build metadata, the RHEL version
portion of the stream name can only be inferred from the RHCOS version
string. This may be cause issues when there is a three digit RHEL
version.