mantle/platform: allow `systemctl is-system-running` nonzero exit codes
The `CheckMachine()` call tolerates `systemctl is-system-running`
returning e.g. `starting` and `initializing` in addition to the usual
`running`. But by design, the command returns a nonzero exit code in
those cases.

Since we're explicitly checking the output string against the acceptable
states, just ignore the command exit code so that we don't fail if the
system is still starting.

Prep for adding a test in which we expect the system to still be
`starting`.