mantle/kola: allow rerun success on timeout
This will allow tests that get aborted due to a timeout to not fail
the overall run if a rerun is attempted and it passes. One example of
where this is useful is if a test times out on initialization and never
reaches the machine via SSH (which we see regularly). In that case
is the failure an issue with the platform bringing up the machine or
a fundamental issue with the software inside FCOS/RHCOS? If it's an
issue with the platform and the rerun succeeds then we don't want
to see a failure. If it's a fundamental issue with the software
inside (i.e. Ignition fails) then the rerun will fail anyway and this
will do no harm.