mantle/kola: allow warnOnly console checks; refactor implementation
This patch enables us to have console checks that are non-fatal and
will print a log message to the screen and nothing more, which was
stated as desirable in [1].

This commit also re-works the implementation of the console/journal
checks in runTest() to deduplicate code. It has the side-effect of
making SkipConsoleWarnings apply to journal checks too, but I think
that's actually a benefit and not a negative.

[1] https://github.com/coreos/coreos-assembler/pull/3450#issuecomment-1549977236