mantle/kola: enhance allTestsDenyListed detection
With 8166991 we would inadvertently hit the "There are no tests to run
because all tests are denylisted" case because we were passing `tests`
to `allTestsDenyListed()`, but that didn't make sense because
`len(tests) == 0`.

Here we make sure to pass all tests that matched patterns to
`allTestsDenyListed()` so we can check against all pattern matching
tests.