mantle/kola: rework denylist error/non-error cases again
It's almost impossible to figure out if a test is denylisted or just
filtered out via arch/platform/tag exclusions. Let's break the
denylist processing to a separate function. Then we can do checks
before and after the denylist processing happens and we'll know
if the tests set is empty because of denylist processing or not.