mantle/platform: Simplify failed unit detection
Do not special case the failed unit detection logic for each
distributions.

If we need to ignore some unit failures in the future for some
distributions, we should add support for this using a proper
kola-denylist entry instead so that we can track it properly.

This removes the special case for RHCOS which is no longer relevant as
we've moved to RHEL 9.

Also remove the test that are not relevant anymore.

This mostly reverts: https://github.com/coreos/coreos-assembler/pull/2261