cosalib: fix member filename when doing our own compression
gzip and zip both embed the input filename in the output archive, and
we're currently including the compression extension (.gz or .zip) in that
name.  For gzip this is harmless, since gunzip ignores it unless -N is
specified.  For zip (currently used only in hyperv), the member filename
is not ignored during decompression, so the decompressed output claims to
be another ZIP file but isn't.

Fix this by stripping the last filename extension from the uncompressed
file before compressing.