mantle/kola/testiso: support testing coreos.liveiso.fromram installs
In https://github.com/coreos/fedora-coreos-config/pull/2544 we added
a new `coreos.liveiso.fromram` kernel argument to instruct the ISO
image to boot completely from memory in order to facilitate installs
back to the same disk the ISO was booted from in iso-as-disk mode
(i.e. dd the ISO file to a hard drive or partition on a hard drive).

Let's test here that we can boot completely from memory and that
/run/media/iso is not mounted.

Currently this doesn't actually do an install to the same disk that
we booted from (actually the iso-as-disk tests don't actually do
installs at all, but just verify they can boot), but rather just
verifies the ISO isn't mounted after the live system boots up, which
should be sufficient.