kola: Add `--qemu-bind-ro`
I'm trying to debug
`kola run --ssh-on-test-failure ext.rpm-ostree.destructive.container-image`
interactively but one problem is that because the test is classified
as disconnected, we run qemu with no networking.  This makes
it impossible to get remote debug tools.

With this I can `--qemu-bind-ro /usr` to get access to the debug
tools in my dev container's `/usr`.  But one needs to manually mount
inside the VM, because injecting the mount into the Ignition
config in kola is a bit trickier.

For reference it looks like e.g.
`mount -t 9p -o trans=virtio,version=9p2000.L /kola/usr /run/hostusr`
in the ssh shell.