live-iso: Write kernel (and hmac) to /boot too
This is useful in order to make the dracut FIPS module work.
For ostree, we put the kernel stuff in `/boot/ostree` in order
to namespace things.  But non-ostree systems tend to use `/boot`
directly, and that's what the dracut module is hardcoded to do.

Now we did add some logic in the dracut module which scrapes the
grub-injected `BOOT_IMAGE`, except this doesn't work for how
we generate the CoreOS Live ISO today because the filenames
of the kernel don't include versions etc.

In the case of the Live ISO though we don't really need to
"ostree namespace" things, so just hardlink the kernel binary
and the hmac file into the traditional places in `/boot`.

We *could* not do this, and do it in our dracut module
in the initramfs in fips mode only for the Live ISO, but that'd
be way more ugly.

xref https://issues.redhat.com/browse/OCPBUGS-15843