qemu: add `env COSA_TEST_CDROM_UNPLUG` to test cdrom unplug
The context is to hacky code to test https://github.com/openshift/os/pull/1346

According to Colin's pointer
https://github.com/openshift/os/pull/1350#issuecomment-1708486101

The cdrom should be `scsi-cd`, as `ide` device can not be
hotplugging. So should not pass `-cdrom` to qemu as it is `ide`
by default which could not work.
```
[coreos-assembler]$ env COSA_TEST_CDROM_UNPLUG=5s cosa run qemu.qcow2 --debug -- \
-drive id=cdrom0,if=none,readonly=on,file=/srv/test.iso \
-device virtio-scsi-pci,id=scsi0 \
-device scsi-cd,bus=scsi0.0,drive=cdrom0
...
2023-09-12T10:54:08Z platform: get cdrom id /machine/peripheral-anon/device[2]
[STARTUP] '2023-09-12T10:54:13Z platform: delete cdrom
...
[core@cosa-devsh ~]$ ls /dev/sr0
ls: cannot access '/dev/sr0': No such file or directory
```