Make `cosa build ostree` an alias for `cosa build container`
Continue to de-brand/de-focus ostree, and increase the focus
on container technology (for which we happen to use ostree
as a backend today).

We already had a new team member ask "how do I have cosa build
a container" - when it already does that by default now, it's
just not obvious.