Stop bumping aarch64 to 8G automatically
The aarch64 RHEL8 kernel used a page size of 64K. In RHEL9, this was
lowered to the usual 4K and a separate kernel was added for 64K.

For testing, this now means that we no longer need to double the memory
for aarch64 by default like we do for ppc64le. If we write tests that
make use of the 64K kernel, we can have them request more memory as
needed.