kola/testiso: combine `fromram` test with `4k` instead
The `iso-offline-install.$firmware` tests are the most common, "default",
path that users installing CoreOS will use. Let's keep that test
untouched so we make sure we always have coverage for it.

To keep the number of tests the same, combine it with the `4k` test
instead, since it differs from the default path in an orthogonal way to
how `fromram` differs.

Follow-up to 095510153 ("mantle/kola/testiso: support testing
coreos.liveiso.fromram installs").