mantle/kola: don't match on tag OR pattern
If a pattern (or patterns) were provided then we should match other
tests (i.e. by tag).  This was causing all tests to get rerun if a
tag was provided it would match all the tests with that tag and not
the pattern provided that described what tests to rerun.

Fixes https://github.com/coreos/coreos-assembler/issues/3041