mantle/kola: limit testbank when doing a rerun
Today we see an issue when trying to do a rerun when a tag was provided
because of the filtertests function logic where all the tests matching
the tag get rerun, not just the tests that failed. Let's limit the
testsBank when doing the rerun so only failed tests can possibly get
selected/run.

Fixes https://github.com/coreos/coreos-assembler/issues/3041