mantle/qemu: neuter virtiofsd seccomp filtering
By default, `virtiofsd` uses seccomp to allow only some syscalls to be
proxied from the guest. In the theme of `--sandbox=none`, let's also
neuter seccomp filtering for our virtiofs usage; the workloads we run in
the supermin/dev VMs are trusted.

Incidentally, this avoids issues like #3635, where some syscalls
were accidentally missing from the allow list. In this case, new
libostree code[[1]] running in the supermin VM when building the
legacy oscontainer calls out to `fstatfs` over virtiofs, which maps to
the blocked `fstatfs64` syscall on ppc64le. (I've opened an upstream
patch[[2]] to fix this, but we don't strictly need it.)

Closes: #3635

[1]: https://github.com/ostreedev/ostree/commit/ba9c9dedffb953738c609fda6502b9b37ee1b6ba
[2]: https://gitlab.com/virtio-fs/virtiofsd/-/merge_requests/200