cmd/remote-session: use `--init` instead of `dumb-init`
I'd like to reuse the `cosa remote-session` bits to run non-cosa images
on remote builders. Since it's mostly a wrapper around `podman --remote`,
it has very few requirements on the target image.

One of the requirements I'd like to loosen is on `dumb-init`. That one
is easy to drop since podman has built-in support for making pid 1 an
init-like system. Use that instead.

We still need to override the entrypoint though so that we can run sleep
instead of cosa.