kola/qemuexec: add `--netboot` option
For local testing, it's useful to have an easy way to netboot a system.
Enable this by exposing QEMU's builtin support for TFTP serving and
BOOTP option.

For iPXE, one can just directly pass the iPXE script. For PXELINUX/GRUB,
you'll likely want to prepare a netboot directory with your artifacts.

Probably this should be streamlined more in the future, and also deduped
more with the related bits in `metal.go`. But anyway, for now this is
immediately useful in helping to test root on iSCSI locally (via iPXE's
`sanboot` option).