kola/qemuexec: allow changing guest network
This is useful when you're nesting VMs and you want the first VM to
be able to access the host. The default host address that QEMU assigns
(e.g. 10.0.2.2, modifiable via the `host=...` netdev knob) doesn't
always work because it's not actually an IP address owned by the host,
but proxied by QEMU itself.

So the source appears to come from localhost, but in some contexts (e.g.
iSCSI), we need the host and the guest to agree that the same IP refers
to the host.

With this, one can start the first VM as usual (e.g. `cosa run`) and the
second VM within with e.g. `cosa run --usernet-addr 10.0.3.0/24` and be
able to talk back to the outer VM via the valid address 10.0.2.15.

To be clear, this can all be done with passthrough QEMU args, so this is
just about making it more convenient.