tests/test_cosalib_meta.py: don't deepcopy GenericBuildMeta
It seems like jsonschema in Python 3.12 (in Fedora 39) changed some
implementation details to now use a data structure that doesn't support
deep copying:

    TypeError: cannot pickle 'rpds.HashTrieMap' object

Instead of doing a deep copy, just create a new object from scratch with
the same inputs. The `w` variable might've been for that; it's currently
unused and is an exact copy of `m`.