osbuild: add support for metal platform
This reworks the coreos.osbuild.mpp.yaml to use the new kernel-cmdline.bls-append
stage [1], enabling us to share earlier stages among the platform specific pipelines.
We have renamed the pipeline stages to be more appropriate. Now we have

- tree
    - the filesystem tree
- raw-image
    - builds on "tree"
    - a raw disk image with nothing platform specific, like kernel arguments
- raw-metal-image
    - builds on "raw-image"
    - adds kernel arguments specific to `metal` platform
- raw-qemu-image
    - builds on "raw-image"
    - adds kernel arguments specific to `qemu` platform
- metal
    - builds on "raw-metal-image"
    - copies out raw image file from "raw-metal-image" stage
    - no other modifications needed
- qemu
    - builds on "raw-qemu-image"
    - converts raw image file from "raw-qemu-image" into a qcow2

We have also modified cmd-buildextend-metal to support calling
runvm-osbuild for either metal or qemu platform images and also added
support for running with the cache qcow2. The cache qcow will allow us
to call osbuild multiple times, with later invocations building on work
done in previous invocations. For example if we checkpiont the `tree`
and `raw-image` stages (as depicted above) then when we call osbuild to
build the `metal` image it won't have to do the work for that. The
benefits of this will compound when we start to build more disk images
in this way.

[1] https://github.com/osbuild/osbuild/pull/1429