create_disk.sh: opt out of EXT4 orphan_file feature on older targets
Fedora 39 has a newer e2fsprogs whose `mkfs.ext4` now turns on a new
`orphan_file` feature by default. This feature is not supported by the
`e2fsck` currently in RHEL 9.

As mentioned in the comment, ideally we'd use the `mkfs.*` tools
from the target system and not from cosa. But for now, just do some
version testing to know when to opt out to avoid having to add another
`image.yaml` knob for it.

See: https://github.com/coreos/coreos-assembler/pull/3653#issuecomment-1813181723