Add a testiso scenario installing coreOS to an iscsi target then boot it
I reused the same approach as other liveISO tests, wrapping
all the setup in a butane config.
Systemd units do all the steps, taking care of the ordering.
It is more verbose compared to imperatively executing
commands to the VM through SSH but the avaiable tooling to do SSH
commands in testiso is not as complete as in regular kola tests,
so doing SSH would end up in a lot of error handling code.
A future work is to merge the testiso scenarios with regular kola tests.

Extra storage mounted into /var to avoir running out of space in the
liveiso.

When installing to the iscsi target with coreos-install
we pass an ignition config to write to a serial device
when the boot is complete.
This serial device is passed through to the kola
completion serial device, so it completes the test.