Add `kola-junit` script to generate JUnit XML files
There's a lot of friction associated with debugging pipeline kola test
failures. One must scroll a bunch and sift through the output to see
which test(s) failed, and if more information is needed, download the
archived artifacts, extract it, and look for the journal and console
there.

Let's make this easier by outputting JUnit XML files. Jenkins can
natively display test results in that format so it's much easier to see
which tests ran, which ones failed, and their outputs (in the Blue Ocean
view, it's the "Tests" tab which we haven't been using so far).

For space efficiency and to reduce noise, we default to skipping over
passed tests.