osbuild: add in patches for enabling 4k disk images
These are the patches from upstream PR:
Upstream https://github.com/osbuild/osbuild/pull/1461

This commit also reworks a bit the patch_osbuild() function
to make it easier for us to apply patches from git in the future.
In git the code is all in the git repo whereas on a system the files
are in various directories under /usr, so the patches pulled from
git are hard to apply unless you do some musical chairs.