create_disk: add version field to aleph file
In the osbuild aleph file work we decided to use `version:` instead
of `build:` here because our COSA build ID isn't exactly ties to the
version of the ostree commit and is essentially a COSA concept. While
we transition to osbuild let's just add a version field here too in
COSA and lazily set it to buildid, even if it's not 100% accurate.

Also, fix a test that checks the `.build` field from .coreos-aleph-version.json