qemu: Support setting up a virtio console (hvc0)
The main motivation here is debugging issues with podman machine
on Apple Virtualization, which only implements virtio, so we need
to talk to hvc0.

However there's no reason for us not to also do this with qemu;
we already have a few references to hvc0 in our console setup.

Turning this on by default will hence require a PR to f-c-c
to add that.

But for now let's make it easier to test via an env var.