create_disk.sh: stop setting immutable bit on sysroot
Nowadays, the sysroot is mounted read-only anyway, so it seems
unnecessary to also slap on an immutable bit. Note this is distinct from
the deployment root immutable bit.

Prompted by the fact that when moving to osbuild, we don't want to have
to care about this and try to replicate it.