buildextend-metal: rework image size and rootfs size handling
This patch changes the strategy to create the disk initially
with the smaller disk size and then resize the disk at the end
if needed to the larger disk size.

This simplifies things a bit since we don't need to pass down
the rootfs size into create_disk.sh.

This is prep for some later patches that add osbuild support for
dynamically setting the disk size and should simplify things since
for both create_disk.sh and osbuild we won't need to pass in the
rootfs size.