fedmsg-broadcast: allow many basearches in `stream.release` msg
Right now, when a new build is released, we emit the `stream.release`
message once per architecture. But an important part of the release job
is to "integrate" all the different arches into a single release. So it
seems much more appropriate to have it emit a single fedmsg and in it,
we list out all the arches in that release.

Prep for adding a consumer to that message.