create_disk.sh: fix Secure Execution build
In 84ce8dca0, we broke the Secure Execution build because the root
filesystem takes up all the remaining space, which leaves no space for
the verity partitions that need to come after.

Hack around this by making the root partition end 384M before the end
of the disk (that's the sum of the sizes of the verity partitions).
Correspondingly add that amount of space to the initial disk size in
`cmd-buildextend-metal`.

Fixes 84ce8dca0 ("buildextend-metal: rework image size and rootfs size
handling").