osbuild: add patches for cache LRU eviction
This is a PR upstream [1], but I'd really like it now so in local
development the cache doesn't become useless once it feels up
(because the current behavior in OSBuild is never evict anything
so the cache will fill up eventually and then never be used for
anything).

Note this patchset includes one additional patch by myself which
adds a log statement that prints out when an eviction happens. This
is purely for informational purposes so I can know when the cache
LRU eviction mechanism is working.

[1] https://github.com/osbuild/osbuild/pull/1503