cmd-build: add --autolock=VERSION
Currently, `cosa build` will enable autolocking automatically if it
detects that it's completing a multi-arch build on an unlocked stream
(as is the case in RHCOS when e.g. building aarch64 after the x86_64
build).

However, there is another use case for autolocking where we want to
create a new build, possibly on the same arch, but reuse the previous
build as a base lockfile. This comes up in Bodhi testing, where we want
to test the Bodhi update with the only change from a known working build
being the new packages.

For this, add an explicit `--autolock`, because we don't want
`cosa build` to try to auto-detect this.

See also: https://github.com/coreos/fedora-coreos-releng-automation/issues/181