[kola testiso] iscsi - misc fixes to improve reliability
- fix invalid After targets in iscsi services
- rename target container to targetd : Target is a systemd w
ord and seems to interfere with systemd, as running
systemctl stop target shuts the vm down
- add a wait for the targetd container to be ready before configuring the
iscsi target. this should be removed when we ship systemd 254 in rhcos
- Add a healthcheck to the target container and make sure we mark the
systemd service as started only when targetcli is ready to receive
commands. Note this is commented out due to an older systemd version in
rhcos
- update virtserial channel completion channel name to "testisocompletion"
to allow the coreos-test-entered-emergency-target.service to write to
it, as it expects this name.
This will allow correct failure detection rather than waiting for QEMU
timeout.
- add a restart policy on setup-targetcli as it may start too fast so
we allow it to restart before marking it failed. Disabled for now
waiting for a systemd update in rhcos