osbuild: copy in platforms.json into /boot
This is something create_disk.sh was doing that we weren't in the
OSBuild flow today. It will make it so `coreos-installer` can pick
up and use this information when performing installs and should
support the efforts to make the ext.config.files.console-config
test start passing again.

xref https://github.com/coreos/fedora-coreos-tracker/issues/1654